#45 - added features (bbox, limit, datetime)
